Moving more deps around
